Immediate Mapping Feature-Complete
The OSM Immediate Mapping has been completely implemented with a configurable memory cache during
conversion which might need some fine-tuning. Further, the libosmpbfreader has been patched to include
a progress bar in the visitor to see the progress. This is just a callback based on bytes read and
bytes expected. With a local static, we show how this is supposed to be used in osm_immediate.cpp.

Let me transform the planet as the next step.